+++ std/pkg/gandora_std/keyword.py
@@ -7,7 +7,7 @@ class GanMatchError(Exception):
     pass
 
 
-def get(*_gan_args) -> object:
+def get(kw: list[tuple[str, object]], key: str, default: object = None) -> object:
     """The first value for `key`, or nil.
 
 ## Parameters
@@ -20,13 +20,7 @@ def get(*_gan_args) -> object:
     1
 """
     while True:
-        match _gan_args:
-            case (kw, key, default,):
-                return (next((v for k, v in (kw) if k == (key)), (default)))
-            case (kw, key,):
-                _gan_args = (kw, key, None)
-                continue
-        raise GanMatchError("no clause of get/2,3 matched " + repr(_gan_args))
+        return (next((v for k, v in (kw) if k == (key)), (default)))
 
 
 def put(kw: list[tuple[str, object]], key: str, value: object) -> list[tuple[str, object]]: